Rename and organize tests
