+++ tests/test_codes.py
@@ -1,0 +1,43 @@
+#    Copyright 2021 SECTRA AB
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+import unittest
+from typing import Type
+
+import pytest
+from parameterized import parameterized
+
+from wsidicom.conceptcode import CidConceptCode, Code
+
+
+@pytest.mark.unittest
+class WsiDicomCodeTests(unittest.TestCase):
+    @parameterized.expand(
+        (
+            code_class,
+            code,
+        )
+        for code_class in CidConceptCode.__subclasses__()
+        for code in code_class.cid.values()
+    )
+    def test_create_code_from_meaning(
+        self, code_class: Type[CidConceptCode], code: Code
+    ):
+        # Arrange
+
+        # Act
+        created_code = code_class(code.meaning)
+
+        # Assert
+        self.assertEqual(code, created_code)
